+++ app/src/main/java/com/example/weatherapp/data/location/Location.kt
@@ -1,9 +1,9 @@
 package com.example.weatherapp.data.location
 
-data class Location(val latitude: Float, val longitude: Float) {
+data class Location(val latitude: Double, val longitude: Double) {
 
   companion object {
-    val ZERO = Location(0.0F, 0.0F)
+    val ZERO = Location(0.0, 0.0)
   }
 }
